fix conversation template
