+++ fastchat/model/model_adapter.py
@@ -15,8 +15,12 @@ from transformers import (
     T5Tokenizer,
 )
 
-
 from fastchat.conversation import Conversation, get_conv_template
+from fastchat.model.compression import load_compress_model
+from fastchat.model.monkey_patch_non_inplace import (
+    replace_llama_attn_with_non_inplace_operations,
+)
+from fastchat.utils import get_gpu_memory
 
 
 class BaseAdapter:
