+++ fastchat/model/monkey_patch_non_inplace.py
